+++ deployment/build-open-source-dist.sh
@@ -25,7 +25,7 @@ source_dir="$source_template_dir/../source"
 echo "------------------------------------------------------------------------------"
 echo "Building open-source folder"
 echo "------------------------------------------------------------------------------"
-[ -e $dist_dir ] && rm -rv $dist_dir
+[ -e $dist_dir ] && rm -rvf $dist_dir
 rm -rf $dist_dir
 mkdir -p $dist_dir
 mkdir -p $dist_template_dir
@@ -34,35 +34,38 @@ echo "--------------------------------------------------------------------------
 echo "Copying Deployment Folder"
 echo "------------------------------------------------------------------------------"
 cp -v $source_template_dir/live-streaming-on-aws.yaml $dist_template_dir
-cp -v $source_template_dir/live-streaming-on-aws-demo.yaml $dist_template_dir
 cp -v $source_template_dir/build-s3-dist.sh $dist_template_dir
 cp -v $source_template_dir/run-unit-tests.sh $dist_template_dir
 
 echo "------------------------------------------------------------------------------"
-echo "Removing Build Files From Source"
-echo "------------------------------------------------------------------------------"
-find $source_dir -iname "node_modules" -type d -exec rm -rv "{}" \; 2> /dev/null
-find $source_dir -type f -name 'package-lock.json' -delete
-
-echo "------------------------------------------------------------------------------"
 echo "Copying Source Folder"
 echo "------------------------------------------------------------------------------"
-cp -rv $source_dir $dist_dir
-cp -v $source_template_dir/architecture.png $dist_dir
-cp -v $source_template_dir/../LICENSE.txt $dist_dir
-cp -v $source_template_dir/../NOTICE.txt $dist_dir
-cp -v $source_template_dir/../README.md $dist_dir
-cp -v $source_template_dir/../CODE_OF_CONDUCT.md $dist_dir
-cp -v $source_template_dir/../CONTRIBUTING.md $dist_dir
-cp -v $source_template_dir/../CHANGELOG.md $dist_dir
+cp -r $source_dir $dist_dir
+cp $source_template_dir/../LICENSE.txt $dist_dir
+cp $source_template_dir/../NOTICE.txt $dist_dir
+cp $source_template_dir/../README.md $dist_dir
+cp $source_template_dir/../CODE_OF_CONDUCT.md $dist_dir
+cp $source_template_dir/../CONTRIBUTING.md $dist_dir
+cp $source_template_dir/../CHANGELOG.md $dist_dir
+
+echo "------------------------------------------------------------------------------"
+echo "Copying Architecture Diagram and GitHub Templates"
+echo "------------------------------------------------------------------------------"
+cp -r $source_template_dir/assets/.github $dist_dir/
+cp $source_template_dir/assets/architecture.png $dist_dir/
+
+echo "------------------------------------------------------------------------------"
+echo "Removing Build Files From Open Source Packaging"
+echo "------------------------------------------------------------------------------"
+find $dist_dir -iname "node_modules" -type d -exec rm -rf "{}" \; 2> /dev/null
+find $dist_dir -iname "build" -type d -exec rm -rf "{}" \; 2> /dev/null
 
 echo "------------------------------------------------------------------------------"
 echo "Creating GitHub zip file"
 echo "------------------------------------------------------------------------------"
 cd $dist_dir
-echo "open source files::"
-ls -lh
-zip -q -r9 ../$1.zip *
-#rm -rf *
+zip -q -r9 ../$1.zip * .github
+rm -rf * 
+rm -rf .github
 mv ../$1.zip .
 echo "Completed building $1.zip"